Prototype of Checkbox Selection Menu
I am trying to make this menu button and its pop up using another scene. My hope is to either: have the popup not change the sidebar autosize, or to align a sneaky second copy of the menu button over the button for the same no-auto-resize-effect. Making a save for the safety of the Editor scene.