Checkbox Menu Status selection working
Got to use a button group for the status buttons. I will work on the ability to delete profiles once I can save them in the first place. The profile name wraps on the checkbox menu, to prevent weird resizing. The menu looks and functions great, I am calling it done for its Minimum release.  When the new profile menu is opened, the picker is set to a random color which I just adore.