First stab at checkbox filling
I tried to make a custom draw for the treeitems. I actually succeeded in making the half/half colored rectangles on the clicked cells. Creating a new checkbox after changing profile turns all existing checkboxes to the new color, whoops. Still looks the part. May be the beginning of the end of the Tree and Treeitems.