Task Spreadsheet Data connected and tested
The gridcontainer now runs an initialization test if no data is loaded into it. GridContainer makes the TaskSpreadsheetData which in turn makes the TaskData(and their checkboxes). Spreadsheet data also calls the taskdata print. Different defaults at different stages tell me if the data is loading properly. Also learned the importance of the separate parameter variable in the init.