Data Manager prototyped
Moving towards managing files and csv in the app. The manager is a separate scene communicating with a signal singleton. Should be able to use the "popup" to make its own screen for main menu.  And processing csv myself means no more addon for it. Useful as a tool, but not what I needed. Plus more hands on learning.