TaskGridContainer gets some love, error.keys error
GridContainer lost its initial test, now has text outputs for the various states. GridContainer now grabs the global tasksheet on _ready and when the global is changed.  Added a MANUALLY TYPED array because Error.keys() is somehow not allowed. But I did get to add a funny "Highscore" feature to the error outputs.