Existing Tasksheet data loads properly!
The loader checks for files (if the folder exists) and passes them to make buttons.  They all get loaded in order, so the "last" data will be the one to load on opening. The loaded resources call init and print their default. I am worried that the data is going wrong, but so far all other references seem fine. Oddity or bug, we will find out in time.