Still approaching Grid
Removed the init function of the resource files. Buttons and modes that old TableTree used to handle are being moved. Editor now handles the buttons, and the mode info is passed to the DataGlobal. Sanitized the input so that resource file names are now saved as snakecase. That ol chestnut. I made a fantabulous button-message function to warn users of requirements. Ever closer to Task Grid without an actual grid. Plus its 5am and the brain just gave out. Really hope to have it out BEFORE this weekend.