Improved and decoupled Editor state buttons
I liked the button_message function so much I want to use it everywhere, so instead of pasting it the third time I moved it to global-autoload! I was about to pass a new parameter for Parent node when I realized it mattered not for the throw away timer. So I let the existing button parameter be the parent. The switch / toggle buttons of Editor are now in the editor.gd and save states to global.  Less parameter relays. Improved switchers' output messages.