More cells
Added in multi-line cells (which will need more love later) and dropdown cells. I tried a few ways to set and update the Arrays for the dropdowns, but they refused to be changed properly. I would run .capitalize on the array and it would print right, but the next use was back to ALLCAP.  Does not seem like consistent behavior.