Checkboxes added!
The checkboxes seem to pop up and populate. Newly created tasks use the color of the current user. AND EXISTING TASKS DON'T CHANGE COLOR when using a different user. That aught to teach Tree/item. It even saves the new tasks properly.  There is no month category for months to sort by so I have to fix that. Otherwise seems great.