Profiles integrated into TaskSheetData
The biggest change was removing the temp profiles from DataGlobal and adding them to TaskSheetData. I made some corrections to the border _draw, I had not adjusted the rectangle for the width. The NewProfileButton is now disabled before data is loaded. The NewProfileButton now has defences against empty names and the color white (which is reserved for default). The random color was so fun I added it as a feature, a smol button. The save-guard button only works to save if the mode is properly triggered. If a change is made but nothing triggers the mode, changes could be lost, so now the save button in the menu can save anytime.  Task settings are introduced, but only added temp debug as functional buttons.