Dropdown rebuild
Went to add profiles to the dropdown cells, and it did not go easy. Mostly because of older code and learning mistakes/confusion. New dropdowns handle the old enums, now they also handle profiles. There is a signal to update the profiles, waiting for use.
Selecting on the dropdown changes the text, sets the stored data, and triggers the save warning. These dropdowns are some of the wierdest ins and outs of UI for cells. ItemsTree would win in still around, popup menus and the like are in 2nd but do not need to be updated much if at all. So the hardest cell to create is now working. New dropdown types SHOULD be easy to add, with an outside updater for the dropdown list.