Task Delete Cell up and running
I added a new cell into the GridContainer. Things behaved and did not explode!  The button includes the task name, so you know what you are clicking on. A button_message provides click-safety and resets the button text. The data and cells are removed, but changes to TaskSheetData  have to be saved or the data will come back. I had not planned on the "soft delete" but works nicely as a failsafe.