Single Task Checkbox reset added!
Added a button at the end of the checkboxes, to reset just that task row. If on daily or weekly, only one month is affected (the current month, also listed in the header). A new signal for the smallest GridContainer reset yet. The other stages of resets were redone as a nice nested flow.