Main Settings first pass
The main settings, mostly for display, are beginning to work. I could have made some earlier saves, but it was not entirely in a working state. Still have some ironing out to do, and most settings will be over run by the Fullscreen setting.  It will be fun and functional, and easy to reuse in future projects.